requirements.txt
bokeh==2.3.2
colour==0.1.5
flameplot==0.1.3
imageio==2.9.0
Keras==2.4.3
lifelines==0.26.0
matplotlib==3.4.2
missingno==0.4.2
mmae==0.2.2
nltk==3.6.2
numba==0.53.1
numpy==1.23.5
pandas==1.5.3
PhenoGraph==1.5.7
plotly==5.9.0
PyYAML==5.4.1
regex==2021.4.4
scikit-learn==1.0.2
scipy==1.10.1
seaborn==0.11.1
shap==0.39.0
six==1.15.0
spacy==3.0.6
statsmodels==0.13.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
textblob==0.15.3
umap-learn==0.5.1
xgboost==1.4.2
xlrd==1.2.0
yellowbrick==1.3.post1
